+++ tests/testthat/test-mt_movebank_visible.R
@@ -12,7 +12,7 @@ test_that("visible testing", {
     TRUE, TRUE, FALSE, TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, TRUE, TRUE, FALSE, TRUE, TRUE
   )
   expect_identical(mt_movebank_visible(d), exp)
-
+  expect_identical(mt_filter_movebank_visible(d), d[exp, ])
   expect_identical(mt_movebank_visible(cbind(visible = rev(exp), d)), rev(exp))
   names(d) <- gsub("-", "_", names(d))
   expect_identical(mt_movebank_visible(d), exp)
@@ -34,5 +34,16 @@ test_that("visible testing", {
   expect_error(
     mt_movebank_visible(d |> dplyr::mutate("manually_marked_valid" = "h")),
     "The values in the `manually_marked_valid` column are expected to be logical"
+  )
+  expect_error(
+    mt_movebank_visible(d |> dplyr::mutate("algorithm_marked_outlier" = as.character(algorithm_marked_outlier))),
+    "The values in the `algorithm_marked_outlier` column are expected to be logical"
+  )
+  expect_error(
+    mt_movebank_visible(d |> dplyr::mutate(
+      "algorithm_marked_outlier" = as.character(algorithm_marked_outlier),
+      "visible" = as.character(algorithm_marked_outlier)
+    )),
+    "The values in the `visible` column are expected to be logical"
   )
 })